% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbStandardize.r
\name{xgbStandardize}
\alias{xgbStandardize}
\title{Standardize the format of XGBoost}
\usage{
xgbStandardize(formula, data, .round = 0, ...)
}
\arguments{
\item{formula}{Model formula used to learn or predict the model}

\item{data}{Input to the model in a data frame format}
}
\value{
A model with the list values of xgboost with new values for
predictor and predictor.factors
}
\description{
Call the xgbStandardize function in lieu of the xgboost function
to be able to use the xgboost function by using a formula and data
frame to learn a model, to use the gridCV function in the buildmodels
package and to make predictions.
}
\details{
If numeric values are passed as the dependent variable in the formula,
it is converted to a factor rounded to the nearest integer. The rounding
can be changed by using the @param .round parameter and passing to it the
digits from the @seealso \code{\link[base:Round]{base::round()}} function.

The xgbStandardize also overrides the behavior of the prediction function
for an xgb.standard model. This allows for the pass through of data frame
as the data object, and provides the prediction probabilities if
decision.value == TRUE.
}
\examples{
model.xgb <- xgbStardize(status~., data, nrounds = 100, eta = 0.01)
pred.xgb <- predict(model.xgb, newdata=data)
prob.xgb <- predict(model.xgb, newdata=data, decision.values=TRUE)
}
\seealso{
\code{\link[xgboost:xgb.train]{xgboost::xgboost()}}
}
\keyword{xgboost}
